refactors region select and base select
