+++ packages/manager/src/components/EnhancedSelect/index.tsx
@@ -1,4 +1,8 @@
-import EnhancedSelect, { Item as _Item } from './Select';
+import EnhancedSelect, {
+  BaseSelectProps as _BaseSelectProps,
+  Item as _Item
+} from './Select';
 /* tslint:disable */
 export interface Item extends _Item {}
+export interface BaseSelectProps extends _BaseSelectProps {}
 export default EnhancedSelect;
