+++ packages/manager/src/components/EnhancedSelect/variants/RegionSelect/index.ts
@@ -1,0 +1,6 @@
+import { ExtendedRegion as _ExtendedRegion } from './RegionSelect';
+
+/* tslint:disable */
+export interface ExtendedRegion extends _ExtendedRegion {}
+
+export { flags, default } from './RegionSelect';
