+++ packages/manager/src/components/SelectRegionPanel/SelectRegionPanel.tsx
@@ -1,11 +1,3 @@
-import CA from 'flag-icon-css/flags/4x3/ca.svg';
-import DE from 'flag-icon-css/flags/4x3/de.svg';
-import UK from 'flag-icon-css/flags/4x3/gb.svg';
-import IN from 'flag-icon-css/flags/4x3/in.svg';
-import JP from 'flag-icon-css/flags/4x3/jp.svg';
-import SG from 'flag-icon-css/flags/4x3/sg.svg';
-import US from 'flag-icon-css/flags/4x3/us.svg';
-import { groupBy, pathOr } from 'ramda';
 import * as React from 'react';
 import { compose } from 'recompose';
 import Paper from 'src/components/core/Paper';
@@ -16,35 +8,11 @@ import {
   WithStyles
 } from 'src/components/core/styles';
 import Typography from 'src/components/core/Typography';
-import Select, { GroupType } from 'src/components/EnhancedSelect/Select';
 import RenderGuard, { RenderGuardProps } from 'src/components/RenderGuard';
 
-import RegionOption, { RegionItem } from './RegionOption';
-
-export const flags = {
-  us: () => <US width="32" height="24" viewBox="0 0 720 480" />,
-  sg: () => <SG width="32" height="24" viewBox="0 0 640 480" />,
-  jp: () => (
-    <JP
-      width="32"
-      height="24"
-      viewBox="0 0 640 480"
-      style={{ backgroundColor: '#fff' }}
-    />
-  ),
-  uk: () => <UK width="32" height="24" viewBox="0 0 640 480" />,
-  de: () => <DE width="32" height="24" viewBox="0 0 720 480" />,
-  ca: () => <CA width="32" height="24" viewBox="0 0 640 480" />,
-  in: () => <IN width="32" height="24" viewBox="0 0 640 480" />
-};
-
-export const selectStyles = {
-  menuList: (base: any) => ({ ...base, maxHeight: `60vh !important` })
-};
-
-export interface ExtendedRegion extends Linode.Region {
-  display: string;
-}
+import RegionSelect, {
+  ExtendedRegion
+} from 'src/components/EnhancedSelect/variants/RegionSelect';
 
 type ClassNames = 'root';
 
@@ -71,79 +39,6 @@ interface Props {
   disabled?: boolean;
 }
 
-export const getRegionOptions = (regions: ExtendedRegion[]) => {
-  const groupedRegions = groupBy<ExtendedRegion>(thisRegion => {
-    if (thisRegion.country.match(/(us|ca)/)) {
-      return 'North America';
-    }
-    if (thisRegion.country.match(/(de|uk)/)) {
-      return 'Europe';
-    }
-    if (thisRegion.country.match(/(jp|sg|in)/)) {
-      return 'Asia';
-    }
-    return 'Other';
-  }, regions);
-  return ['North America', 'Europe', 'Asia', 'Other'].reduce(
-    (accum, thisGroup) => {
-      if (
-        !groupedRegions[thisGroup] ||
-        groupedRegions[thisGroup].length === 0
-      ) {
-        return accum;
-      }
-      return [
-        ...accum,
-        {
-          label: thisGroup,
-          options: groupedRegions[thisGroup]
-            .map(thisRegion => ({
-              label: thisRegion.display,
-              value: thisRegion.id,
-              flag: pathOr(
-                () => null,
-                [thisRegion.country.toLocaleLowerCase()],
-                flags
-              ),
-              country: thisRegion.country
-            }))
-            .sort(sortRegions)
-        }
-      ];
-    },
-    []
-  );
-};
-
-export const getSelectedRegionById = (
-  regionID: string,
-  options: GroupType[]
-) => {
-  const regions = options.reduce(
-    (accum, thisGroup) => [...accum, ...thisGroup.options],
-    []
-  );
-  return regions.find(thisRegion => regionID === thisRegion.value);
-};
-
-const sortRegions = (region1: RegionItem, region2: RegionItem) => {
-  // By country desc so USA is on top
-  if (region1.country > region2.country) {
-    return -1;
-  }
-  if (region1.country < region2.country) {
-    return 1;
-  }
-  // Alphabetically by display name, which is the city
-  if (region1.label < region2.label) {
-    return -1;
-  }
-  if (region1.label > region2.label) {
-    return 1;
-  }
-  return 0;
-};
-
 const SelectRegionPanel: React.FC<Props & WithStyles<ClassNames>> = props => {
   const {
     classes,
@@ -158,8 +53,6 @@ const SelectRegionPanel: React.FC<Props & WithStyles<ClassNames>> = props => {
   if (props.regions.length === 0) {
     return null;
   }
-
-  const options = getRegionOptions(regions);
 
   return (
     <>
@@ -178,17 +71,12 @@ const SelectRegionPanel: React.FC<Props & WithStyles<ClassNames>> = props => {
             to find the best region for your current location.
           </Typography>
         )}
-        <Select
-          isClearable={false}
-          value={getSelectedRegionById(selectedID || '', options)}
-          label="Select a region"
+        <RegionSelect
           errorText={error}
           disabled={disabled}
-          placeholder="Regions"
-          options={options}
-          onChange={(selection: RegionItem) => handleSelection(selection.value)}
-          components={{ Option: RegionOption }}
-          styleOverrides={selectStyles}
+          handleSelection={handleSelection}
+          regions={regions}
+          selectedID={selectedID}
         />
       </Paper>
     </>
