+++ packages/manager/src/components/SelectRegionPanel/index.ts
@@ -1,10 +1,1 @@
-import SelectRegionPanel, {
-  ExtendedRegion as _ExtendedRegion,
-  flags
-} from './SelectRegionPanel';
-/* tslint:disable */
-export interface ExtendedRegion extends _ExtendedRegion {}
-
-export { flags };
-
-export default SelectRegionPanel;
+export { default } from './SelectRegionPanel';
